more score saving stuff

+++ Assets/Scripts/GameManager.cs
@@ -5,6 +5,7 @@ using DedicatedServer;
 using TMPro;
 using Newtonsoft.Json;
 using System.IO;
+using System;
 
 public class GameManager : MonoBehaviour
 {
@@ -13,20 +14,16 @@ public class GameManager : MonoBehaviour
     public TMP_Text pointsText;
 
 
-    int[] scores = new int[]
-    {
-        8384,
-        234234,
-        2343,
-        3754
-    };
+    List<int> scores = new List<int>();
+
+    int[] _scores;
 
     void Start()
     {
         Debug.Log(Application.persistentDataPath);
         UpdatePointsText();
         TestPoints();
-        SaveScores();
+        SaveScoresToDisk();
     }
 
     // Update is called once per frame
@@ -49,16 +46,57 @@ public class GameManager : MonoBehaviour
 
         for (int i = 0; i < timesToAddPoints; i++)
         {
-            AddPoints(Random.Range(0, 2000));
+            AddPoints(UnityEngine.Random.Range(0, 2000));
         }
     }
-    public void SaveScores()
+    public void SaveScore(int score)
     {
-        string sussy = JsonConvert.SerializeObject(scores);
-        Debug.Log(sussy);
+        _scores = PlayerPrefsX.GetIntArray("Scores");
+        scores.AddRange(_scores);
+        scores.Add(score);
+        print("Score added to list: " + score);
+        SaveScoresToDisk();
+    }
+    public void SaveScoresToDisk()
+    {
+        // convert list of scores to array
+        _scores = scores.ToArray();
+        PlayerPrefsX.SetIntArray("Scores", _scores);
 
+        // print stuff
+        print("Score array added to PlayerPrefs");
+
+        // sort scores
+        Array.Sort(_scores);
+
+        // convert to json data
+        string sussy = JsonConvert.SerializeObject(scores);
+
+        // print more stuff
+        print(sussy);
+
+        // generate filepath for scores
         string path = Path.Combine(Application.persistentDataPath, "scores.json");
 
-        File.WriteAllText(path, sussy);
+        // check if file exists already then generate the json
+        if(!File.Exists(path))
+        {
+            File.WriteAllText(path, sussy);
+
+            // print even more stuff
+            print("Successfully wrote score to disk");
+        }
+        else
+        {
+            File.Delete(path);
+            File.WriteAllText(path, sussy);
+
+            // how much stuff we gonna print
+            print("Successfully wrote score to disk");
+        }
+    }
+    public void EndGame()
+    {
+        SaveScore(points);
     }
 }
